+++ lib/solr.js
@@ -11,15 +11,16 @@
  */
 
 var http = require('http'),
-   https = require('https'),
-   Query = require('./query'),
-   querystring = require('querystring'),
-   format = require('./utils/format'),
-   SolrError = require('./error/solr-error'),
-   JSONStream = require('JSONStream'),
-   duplexer = require('duplexer'),
-   request = require('request'),
-   JSONbig = require('json-bigint');
+  https = require('https'),
+  Pool = require('poolee'),
+  Query = require('./query'),
+  querystring = require('querystring'),
+  format = require('./utils/format'),
+  SolrError = require('./error/solr-error'),
+  JSONStream = require('JSONStream'),
+  duplexer = require('duplexer'),
+  request = require('request'),
+  JSONbig = require('json-bigint');
 
 /**
  * Expose `createClient()`.
@@ -43,16 +44,15 @@ exports.createClient = createClient;
  * @api public
  */
 
-function createClient(host, port, core, path, agent, secure, bigint){
-  var options = (typeof host === 'object') ? host : {
-      host : host,
-      port : port,
-      core : core,
-      path : path,
-      agent : agent,
-      secure : secure,
-      bigint : bigint
-   };
+function createClient(endpoints, core, path, agent, secure, bigint) {
+  var options = (typeof endpoints === 'object') ? endpoints : {
+    endpoints: endpoints,
+    core: core,
+    path: path,
+    agent: agent,
+    secure: secure,
+    bigint: bigint
+  };
   return new Client(options);
 }
 
@@ -74,24 +74,24 @@ function createClient(host, port, core, path, agent, secure, bigint){
  * @api private
  */
 
-function Client(options){
-   this.options = {
-      host : options.host || '127.0.0.1',
-      port : options.port || '8983',
-      core : options.core || '',
-      path : options.path || '/solr',
-      agent : options.agent,
-      secure : options.secure || false,
-      bigint : options.bigint || false
-   };
+function Client(options) {
+  this.options = {
+    endpoints: options.endpoints || ['127.0.0.1:8983'],
+    core: options.core || '',
+    path: options.path || '/solr',
+    agent: options.agent,
+    bigint: options.bigint || false
+  };
+  var protocol = options.secure ? https : http;
+  this.pool = new Pool(protocol, this.options.endpoints);
 
-   // Default paths of all request handlers
-   this.UPDATE_JSON_HANDLER = 'update/json';
-   this.UPDATE_HANDLER = 'update';
-   this.SELECT_HANDLER = 'select';
-   this.ADMIN_PING_HANDLER = 'admin/ping';
-   this.REAL_TIME_GET_HANDLER = 'get';
-   this.SPELL_HANDLER = 'spell';
+  // Default paths of all request handlers
+  this.UPDATE_JSON_HANDLER = 'update/json';
+  this.UPDATE_HANDLER = 'update';
+  this.SELECT_HANDLER = 'select';
+  this.ADMIN_PING_HANDLER = 'admin/ping';
+  this.REAL_TIME_GET_HANDLER = 'get';
+  this.SPELL_HANDLER = 'spell';
 }
 
 /**
@@ -104,11 +104,11 @@ function Client(options){
  * @api public
  */
 
-Client.prototype.basicAuth = function(username,password){
-   var self = this;
-   this.options.authorization = 'Basic ' + new Buffer(username + ':' + password).toString('base64');
-   return self;
-}
+Client.prototype.basicAuth = function(username, password) {
+  var self = this;
+  this.options.authorization = 'Basic ' + new Buffer(username + ':' + password).toString('base64');
+  return self;
+};
 
 /**
  * Remove authorization header
@@ -117,11 +117,11 @@ Client.prototype.basicAuth = function(username,password){
  * @api public
  */
 
-Client.prototype.unauth = function(){
-   var self = this;
-   delete this.options.authorization;
-   return self;
-}
+Client.prototype.unauth = function() {
+  var self = this;
+  delete this.options.authorization;
+  return self;
+};
 
 /**
  * Add a document or a list of documents
@@ -136,15 +136,15 @@ Client.prototype.unauth = function(){
  * @api public
  */
 
-Client.prototype.add = function(docs,options,callback){
-   if(typeof(options) === 'function'){
-      callback = options;
-      options = {};
-   }
-   docs = format.dateISOify(docs); // format `Date` object into string understable for Solr as a date.
-   docs = Array.isArray(docs) ? docs : [docs];
-   return this.update(docs,options,callback);
-}
+Client.prototype.add = function(docs, options, callback) {
+  if (typeof(options) === 'function') {
+    callback = options;
+    options = {};
+  }
+  docs = format.dateISOify(docs); // format `Date` object into string understable for Solr as a date.
+  docs = Array.isArray(docs) ? docs : [docs];
+  return this.update(docs, options, callback);
+};
 
 /**
  * Get a document by id or a list of documents by ids using the Real-time-get feature
@@ -160,16 +160,16 @@ Client.prototype.add = function(docs,options,callback){
  * @api public
  */
 
-Client.prototype.realTimeGet = function(ids, query, callback){
-   if(typeof query === 'function'){
-      callback = query;
-      query = {};
-   }
-   ids = Array.isArray(ids) ? ids : [ids];
-   query.ids = ids.join(',');
+Client.prototype.realTimeGet = function(ids, query, callback) {
+  if (typeof query === 'function') {
+    callback = query;
+    query = {};
+  }
+  ids = Array.isArray(ids) ? ids : [ids];
+  query.ids = ids.join(',');
 
-   return this.get(this.REAL_TIME_GET_HANDLER,query,callback);
-}
+  return this.get(this.REAL_TIME_GET_HANDLER, query, callback);
+};
 
 /**
  * Add the remote resource located at the given path `options.path` into the Solr database.
@@ -187,21 +187,21 @@ Client.prototype.realTimeGet = function(ids, query, callback){
  * @api public
  */
 
-Client.prototype.addRemoteResource = function(options,callback){
-   options.parameters = options.parameters || {};
-   options.format = (options.format === 'xml' ? '' : options.format || ''); // reason: the default route of the XmlUpdateRequestHandle is /update and not /update/xml.
-   options.parameters.commit = (options.parameters.commit === undefined ? false : options.parameters.commit);
-   options.parameters['stream.contentType'] = options.contentType || 'text/plain;charset=utf-8';
-   if(options.path.match(/^https?:\/\//)){
-      options.parameters['stream.url'] = options.path;
-   }else{
-      options.parameters['stream.file'] = options.path;
-   }
+Client.prototype.addRemoteResource = function(options, callback) {
+  options.parameters = options.parameters || {};
+  options.format = (options.format === 'xml' ? '' : options.format || ''); // reason: the default route of the XmlUpdateRequestHandle is /update and not /update/xml.
+  options.parameters.commit = (options.parameters.commit === undefined ? false : options.parameters.commit);
+  options.parameters['stream.contentType'] = options.contentType || 'text/plain;charset=utf-8';
+  if (options.path.match(/^https?:\/\//)) {
+    options.parameters['stream.url'] = options.path;
+  } else {
+    options.parameters['stream.file'] = options.path;
+  }
 
-   var handler = this.UPDATE_HANDLER + '/' + options.format.toLowerCase();
-   var query = querystring.stringify(options.parameters);
-   return this.get(handler,query,callback);
-}
+  var handler = this.UPDATE_HANDLER + '/' + options.format.toLowerCase();
+  var query = querystring.stringify(options.parameters);
+  return this.get(handler, query, callback);
+};
 
 /**
  * Create a writable/readable `Stream` to add documents into the Solr database
@@ -212,31 +212,31 @@ Client.prototype.addRemoteResource = function(options,callback){
  * @api public
  */
 
-Client.prototype.createAddStream = function(options){
-   var path = [this.options.path,this.options.core, this.UPDATE_JSON_HANDLER + '?' + querystring.stringify(options) +'&wt=json']
-      .filter(function(element){
-         return element;
-      })
-      .join('/');
-   var headers = {
-      'content-type' : 'application/json',
-      'charset' : 'utf-8'
-   };
-   if(this.options.authorization){
-      headers['authorization'] = this.options.authorization;
-   }
-   var protocol = this.options.secure ? 'https' : 'http';
-   var optionsRequest = {
-      url : protocol + '://' + this.options.host +':' + this.options.port + path ,
-      method : 'POST',
-      headers : headers
-   };
-   var jsonStreamStringify = JSONStream.stringify();
-   var postRequest = request(optionsRequest);
-   jsonStreamStringify.pipe(postRequest);
-   var duplex = duplexer(jsonStreamStringify,postRequest);
-   return duplex ;
-}
+Client.prototype.createAddStream = function(options) {
+  var path = [this.options.path, this.options.core, this.UPDATE_JSON_HANDLER + '?' + querystring.stringify(options) + '&wt=json']
+    .filter(function(element) {
+      return element;
+    })
+    .join('/');
+  var headers = {
+    'content-type': 'application/json',
+    'charset': 'utf-8'
+  };
+  if (this.options.authorization) {
+    headers.authorization = this.options.authorization;
+  }
+  var protocol = this.options.secure ? 'https' : 'http';
+  var optionsRequest = {
+    url: protocol + '://' + this.options.host + ':' + this.options.port + path,
+    method: 'POST',
+    headers: headers
+  };
+  var jsonStreamStringify = JSONStream.stringify();
+  var postRequest = request(optionsRequest);
+  jsonStreamStringify.pipe(postRequest);
+  var duplex = duplexer(jsonStreamStringify, postRequest);
+  return duplex;
+};
 
 /**
  * Commit last added and removed documents, that means your documents are now indexed.
@@ -250,16 +250,16 @@ Client.prototype.createAddStream = function(options){
  * @api public
  */
 
-Client.prototype.commit = function(options,callback){
-   if(typeof(options) === 'function'){
-      callback = options;
-      options = {};
-   }
-   var data = {
-      commit : options || {}
-   };
-   return this.update(data,callback);
-}
+Client.prototype.commit = function(options, callback) {
+  if (typeof(options) === 'function') {
+    callback = options;
+    options = {};
+  }
+  var data = {
+    commit: options || {}
+  };
+  return this.update(data, callback);
+};
 
 /**
  * Call Lucene's IndexWriter.prepareCommit, the changes won't be visible in the index.
@@ -272,9 +272,11 @@ Client.prototype.commit = function(options,callback){
  * @api public
  */
 
-Client.prototype.prepareCommit = function(callback){
-   return this.update({},{ prepareCommit : true},callback);
-}
+Client.prototype.prepareCommit = function(callback) {
+  return this.update({}, {
+    prepareCommit: true
+  }, callback);
+};
 
 /**
  * Soft commit all changes
@@ -287,9 +289,11 @@ Client.prototype.prepareCommit = function(callback){
  * @api public
  */
 
-Client.prototype.softCommit = function(callback){
-   return this.update({},{ softCommit : true},callback);
-}
+Client.prototype.softCommit = function(callback) {
+  return this.update({}, {
+    softCommit: true
+  }, callback);
+};
 
 /**
  * Delete documents based on the given `field` and `text`.
@@ -305,19 +309,19 @@ Client.prototype.softCommit = function(callback){
  * @api public
  */
 
-Client.prototype.delete = function(field,text,options,callback) {
-   if(typeof(options) === 'function'){
-      callback = options;
-      options = {};
-   }
-   text = format.dateISOify(text);
-   var data = {
-      'delete' :  {
-         query : field +  ':'  + format.escapeSpecialChars(text)
-      }
-   };
-   return this.update(data,options,callback);
-}
+Client.prototype.delete = function(field, text, options, callback) {
+  if (typeof(options) === 'function') {
+    callback = options;
+    options = {};
+  }
+  text = format.dateISOify(text);
+  var data = {
+    'delete': {
+      query: field + ':' + format.escapeSpecialChars(text)
+    }
+  };
+  return this.update(data, options, callback);
+};
 
 /**
  * Delete a range of documents based on the given `field`, `start` and `stop` arguments.
@@ -333,17 +337,17 @@ Client.prototype.delete = function(field,text,options,callback) {
  * @api public
  */
 
-Client.prototype.deleteByRange = function(field,start,stop,options,callback){
-   if(typeof(options) === 'function'){
-      callback = options;
-      options = {};
-   }
-   start = format.dateISOify(start);
-   stop = format.dateISOify(stop);
+Client.prototype.deleteByRange = function(field, start, stop, options, callback) {
+  if (typeof(options) === 'function') {
+    callback = options;
+    options = {};
+  }
+  start = format.dateISOify(start);
+  stop = format.dateISOify(stop);
 
-   var query = field + ':[' + start + ' TO ' + stop + ']';
-   return this.deleteByQuery(query,options,callback);
-}
+  var query = field + ':[' + start + ' TO ' + stop + ']';
+  return this.deleteByQuery(query, options, callback);
+};
 
 /**
  * Delete the document with the given `id`
@@ -358,18 +362,18 @@ Client.prototype.deleteByRange = function(field,start,stop,options,callback){
  * @api public
  */
 
-Client.prototype.deleteByID = function(id,options,callback){
-   if(typeof(options) === 'function'){
-      callback = options;
-      options = {};
-   }
-   var data = {
-      'delete' : {
-         id : id
-      }
-   };
-   return this.update(data,options,callback);
-}
+Client.prototype.deleteByID = function(id, options, callback) {
+  if (typeof(options) === 'function') {
+    callback = options;
+    options = {};
+  }
+  var data = {
+    'delete': {
+      id: id
+    }
+  };
+  return this.update(data, options, callback);
+};
 
 /**
  * Delete documents matching the given `query`
@@ -384,18 +388,18 @@ Client.prototype.deleteByID = function(id,options,callback){
  * @api public
  */
 
-Client.prototype.deleteByQuery = function(query,options,callback){
-   if(typeof(options) === 'function'){
-      callback = options;
-      options = {};
-   }
-   var data = {
-      'delete' : {
-         query : query
-      }
-   };
-   return this.update(data,options,callback);
-}
+Client.prototype.deleteByQuery = function(query, options, callback) {
+  if (typeof(options) === 'function') {
+    callback = options;
+    options = {};
+  }
+  var data = {
+    'delete': {
+      query: query
+    }
+  };
+  return this.update(data, options, callback);
+};
 
 
 /**
@@ -410,9 +414,9 @@ Client.prototype.deleteByQuery = function(query,options,callback){
  * @api public
  */
 
-Client.prototype.deleteAll = function(options,callback){
-   return this.deleteByQuery('*:*',options,callback);
-}
+Client.prototype.deleteAll = function(options, callback) {
+  return this.deleteByQuery('*:*', options, callback);
+};
 
 /**
  * Optimize the index
@@ -426,16 +430,16 @@ Client.prototype.deleteAll = function(options,callback){
  * @api public
  */
 
-Client.prototype.optimize = function(options,callback){
-   if(typeof(options) === 'function'){
-      callback = options;
-      options = {};
-   }
-   var data = {
-      optimize : options || {}
-   };
-   return this.update(data,callback);
-}
+Client.prototype.optimize = function(options, callback) {
+  if (typeof(options) === 'function') {
+    callback = options;
+    options = {};
+  }
+  var data = {
+    optimize: options || {}
+  };
+  return this.update(data, callback);
+};
 
 /**
  * Rollback all add/delete commands made since the last commit.
@@ -448,12 +452,12 @@ Client.prototype.optimize = function(options,callback){
  * @api public
  */
 
-Client.prototype.rollback = function(callback){
-   var data = {
-      rollback : {}
-   };
-   return this.update(data,callback);
-}
+Client.prototype.rollback = function(callback) {
+  var data = {
+    rollback: {}
+  };
+  return this.update(data, callback);
+};
 
 /**
  * Send an update command to the Solr server with the given `data` stringified in the body.
@@ -468,31 +472,29 @@ Client.prototype.rollback = function(callback){
  * @api private
  */
 
-Client.prototype.update = function(data,options,callback){
-   if(typeof(options) === 'function'){
-      callback = options;
-      options = {};
-   }
+Client.prototype.update = function(data, options, callback) {
+  var self = this;
+  if (typeof(options) === 'function') {
+    callback = options;
+    options = {};
+  }
 
-   var json = pickJSON(this.options.bigint).stringify(data);
-   var fullPath = [this.options.path,this.options.core, this.UPDATE_JSON_HANDLER + '?' + querystring.stringify(options) +'&wt=json']
-                              .filter(function(element){
-                                 return element;
-                              })
-                              .join('/');
+  var json = pickJSON(this.options.bigint).stringify(data);
+  var fullPath = [this.options.path, this.options.core, this.UPDATE_JSON_HANDLER + '?' + querystring.stringify(options) + '&wt=json']
+    .filter(function(element) {
+      return element;
+    })
+    .join('/');
 
-   var params = {
-      host : this.options.host,
-      port : this.options.port,
-      fullPath : fullPath,
-      json : json,
-      secure : this.options.secure,
-      bigint : this.options.bigint,
-      authorization : this.options.authorization,
-      agent : this.options.agent
-   };
-   return postJSON(params,callback);
-}
+  var params = {
+    fullPath: fullPath,
+    json: json,
+    bigint: this.options.bigint,
+    authorization: this.options.authorization,
+    agent: this.options.agent
+  };
+  return self.postJSON(params, callback);
+};
 
 /**
  * Search documents matching the `query`
@@ -506,9 +508,9 @@ Client.prototype.update = function(data,options,callback){
  * @api public
  */
 
-Client.prototype.search = function(query,callback){
-   return this.get(this.SELECT_HANDLER, query, callback);
-}
+Client.prototype.search = function(query, callback) {
+  return this.get(this.SELECT_HANDLER, query, callback);
+};
 
 /**
  * Search for all documents
@@ -521,9 +523,9 @@ Client.prototype.search = function(query,callback){
  * @api public
  */
 
-Client.prototype.searchAll = function(callback){
-   return this.search('q=*', callback);
-}
+Client.prototype.searchAll = function(callback) {
+  return this.search('q=*', callback);
+};
 
 /**
  * Search documents matching the `query`
@@ -539,9 +541,9 @@ Client.prototype.searchAll = function(callback){
  * @api public
  */
 
-Client.prototype.spell = function(query,callback){
-   return this.get(this.SPELL_HANDLER, query, callback);
-}
+Client.prototype.spell = function(query, callback) {
+  return this.get(this.SPELL_HANDLER, query, callback);
+};
 
 /**
  * Send an arbitrary HTTP GET request to Solr on the specified `handler` (as Solr like to call it i.e path)
@@ -556,36 +558,34 @@ Client.prototype.spell = function(query,callback){
  * @api public
  */
 
-Client.prototype.get = function(handler,query,callback){
+Client.prototype.get = function(handler, query, callback) {
 
-   var parameters = '';
-   if(typeof query === 'function'){
-      callback = query;
-   }else if(query instanceof Query){
-      parameters += query.build();
-   }else if(typeof query === 'object'){
-      parameters += querystring.stringify(query);
-   }else if(typeof query === 'string'){
-      parameters += query;
-   }
+  var parameters = '',
+    self = this;
+  if (typeof query === 'function') {
+    callback = query;
+  } else if (query instanceof Query) {
+    parameters += query.build();
+  } else if (typeof query === 'object') {
+    parameters += querystring.stringify(query);
+  } else if (typeof query === 'string') {
+    parameters += query;
+  }
 
-   var fullPath = [this.options.path,this.options.core,handler + '?' + parameters + '&wt=json']
-                              .filter(function(element){
-                                 return element;
-                              })
-                              .join('/');
+  var fullPath = [this.options.path, this.options.core, handler + '?' + parameters + '&wt=json']
+    .filter(function(element) {
+      return element;
+    })
+    .join('/');
 
-   var params = {
-      host : this.options.host,
-      port : this.options.port,
-      fullPath : fullPath,
-      secure : this.options.secure,
-      bigint : this.options.bigint,
-      authorization : this.options.authorization,
-      agent : this.options.agent
-   };
-   return getJSON(params,callback);
-}
+  var params = {
+    fullPath: fullPath,
+    bigint: this.options.bigint,
+    authorization: this.options.authorization,
+    agent: this.options.agent
+  };
+  return self.getJSON(params, callback);
+};
 
 /**
  * Create an instance of `Query`
@@ -594,9 +594,9 @@ Client.prototype.get = function(handler,query,callback){
  * @api public
  */
 
-Client.prototype.createQuery = function(){
-   return new Query();
-}
+Client.prototype.createQuery = function() {
+  return new Query();
+};
 
 /**
  * Expose `format.escapeSpecialChars` from `Client.escapeSpecialChars`
@@ -615,20 +615,8 @@ Client.prototype.escapeSpecialChars = format.escapeSpecialChars;
  * @api public
  */
 
-Client.prototype.ping = function(callback){
-   return this.get(this.ADMIN_PING_HANDLER, callback);
-}
-
-/**
- * Pick appropriate protocol based on the given `secure` flag
- *
- * @param {Boolean} secure -
- * @return {Object} http or https module
- * @api private
- */
-
-function pickProtocol(secure){
-   return secure ? https : http;
+Client.prototype.ping = function(callback) {
+  return this.get(this.ADMIN_PING_HANDLER, callback);
 };
 
 
@@ -639,9 +627,9 @@ function pickProtocol(secure){
  * @api private
  */
 
-function pickJSON(bigint){
-   return bigint ? JSONbig : JSON;
-};
+function pickJSON(bigint) {
+  return bigint ? JSONbig : JSON;
+}
 
 /**
  * HTTP POST request. Send update commands to the Solr server (commit, add, delete, optimize)
@@ -664,39 +652,34 @@ function pickJSON(bigint){
  * @api private
  */
 
-function postJSON(params,callback){
-   var headers = {
-      'content-type' : 'application/json; charset=utf-8',
-      'content-length':  Buffer.byteLength(params.json),
-      'accept' : 'application/json; charset=utf-8'
-   };
-   if(params.authorization){
-      headers['authorization'] = params.authorization;
-   }
-   var options = {
-      host : params.host,
-      port : params.port,
-      method : 'POST',
-      headers : headers,
-      path : params.fullPath
-   };
+Client.prototype.postJSON = function(params, callback) {
+  var headers = {
+    'content-type': 'application/json; charset=utf-8',
+    'content-length': Buffer.byteLength(params.json),
+    'accept': 'application/json; charset=utf-8'
+  };
+  if (params.authorization) {
+    headers.authorization = params.authorization;
+  }
+  var options = {
+    host: params.host,
+    port: params.port,
+    method: 'POST',
+    headers: headers,
+    path: params.fullPath
+  };
 
-   if(params.agent !== undefined){
-      options.agent = params.agent;
-   }
+  if (params.agent !== undefined) {
+    options.agent = params.agent;
+  }
 
-   var request = pickProtocol(params.secure).request(options);
-
-   request.on('response', handleJSONResponse(request, params.bigint, callback));
-
-   request.on('error',function onError(err){
-      if (callback) callback(err,null);
-   });
-
-   request.write(params.json);
-   request.end();
-
-   return request;
+  self.pool.request(options, params.json, function(err, response, body) {
+    if (err) {
+      callback(err);
+    } else {
+      handleJSONResponse(response, body, params.bigint, callback);
+    }
+  });
 };
 
 /**
@@ -719,36 +702,36 @@ function postJSON(params,callback){
  * @api private
  */
 
-function getJSON(params,callback){
-   var headers = {
-      'accept' : 'application/json; charset=utf-8'
-   };
-   var options = {
-      host : params.host,
-      port : params.port,
-      path : params.fullPath,
-      headers : headers
-   };
+Client.prototype.getJSON = function(params, callback) {
+  var self = this;
+  var headers = {
+    'accept': 'application/json; charset=utf-8'
+  };
+  var options = {
+    host: params.host,
+    port: params.port,
+    method: 'GET',
+    path: params.fullPath,
+    headers: headers
+  };
 
-   if(params.agent !== undefined){
-      options.agent = params.agent;
-   }
+  if (params.agent !== undefined) {
+    options.agent = params.agent;
+  }
 
-    if(params.authorization){
-      var headers = {
-         'authorization' : params.authorization
-      };
-      options.headers = headers;
-   }
+  if (params.authorization) {
 
-   var request = pickProtocol(params.secure).get(options);
+    headers.authorization = params.authorization;
+    options.headers = headers;
+  }
 
-   request.on('response', handleJSONResponse(request, params.bigint, callback));
-
-   request.on('error',function(err){
-      if (callback) callback(err,null);
-   });
-   return request;
+  var request = self.pool.request(options, function(err, response, body) {
+    if (err) {
+      callback(err, null);
+    } else {
+      handleJSONResponse(response, body, params.bigint, callback);
+    }
+  });
 };
 
 /**
@@ -761,32 +744,22 @@ function getJSON(params,callback){
  * @api private
  */
 
-function handleJSONResponse(request, bigint, callback){
-   return function onJSONResponse(response){
-      var text = '';
-      var err = null;
-      var data = null;
-
-      // This properly handles multi-byte characters
-      response.setEncoding('utf-8');
-
-      response.on('data',function(chunk){
-         text += chunk;
-      });
-
-      response.on('end',function(){
-         if(response.statusCode < 200 || response.statusCode > 299){
-            err = new SolrError(request,response,text);
-            if(callback)  callback(err,null);
-         }else{
-            try{
-               data = pickJSON(bigint).parse(text);
-            }catch(error){
-               err = error;
-            }finally{
-               if(callback)  callback(err,data);
-            }
-         }
-      });
-   };
-};
+function handleJSONResponse(response, body, bigint, callback) {
+  var err = null;
+  var data = null;
+  var request = response.req;
+  // This properly handles multi-byte characters
+  response.setEncoding('utf-8');
+  if (response.statusCode < 200 || response.statusCode > 299) {
+    err = new SolrError(request, response, body);
+    if (callback) callback(err, null);
+  } else {
+    try {
+      data = pickJSON(bigint).parse(body);
+    } catch (error) {
+      err = error;
+    } finally {
+      if (callback) callback(err, data);
+    }
+  }
+}
