+++ test/config.json
@@ -1,6 +1,6 @@
 {
   "client": {
-    "host": "127.0.0.1",
+    "host": ['10.64.30.11', '10.64.30.12'],
     "port": 8983,
     "path": "/solr",
     "core": null
